pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aigp"
version = "2.0.0"
description = "Autonomous Drone Racing Stack for Anduril AI Grand Prix — Ganymede Systems"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Proprietary"}
authors = [
    {name = "Brendan O'Brien", email = "brendanobrien097@gmail.com"},
]

dependencies = [
    "torch>=2.5",
    "numpy>=1.26",
    "scipy>=1.14",
    "pyyaml>=6.0",
    "opencv-python>=4.10",
    "ultralytics>=8.3",
    "filterpy>=1.4",
    "onnxruntime-gpu>=1.20",
]

[project.optional-dependencies]
sim = [
    "isaacsim>=4.5",
    "isaaclab>=2.0",
    "skrl[torch]>=1.4",
    "rsl-rl>=2.3",
]
ros = [
    "px4-msgs",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.8",
    "wandb",
]
all = ["aigp[sim,ros,dev]"]

[project.scripts]
aigp-train = "scripts.train:main"
aigp-evaluate = "scripts.evaluate:main"
aigp-export = "scripts.export:main"
aigp-deploy = "scripts.deploy:main"

[tool.setuptools.packages.find]
include = ["aigp*"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: marks tests requiring Isaac Sim (deselect with '-m \"not integration\"')",
]
